test: parallel jest
