+++ __tests__/integration/gallery.spec.ts
@@ -1,0 +1,12 @@
+import { generateCanvasTestCases } from './utils/generator';
+
+const TEST_CASES = [
+  {
+    name: 'fujian',
+    sleepTime: 500,
+  },
+];
+
+describe('Gallery Snapshot', () => {
+  generateCanvasTestCases('Gallery', TEST_CASES);
+});
