+++ __tests__/integration/mask.spec.ts
@@ -1,0 +1,11 @@
+import { generateCanvasTestCases } from './utils/generator';
+
+const TEST_CASES = [
+  {
+    name: 'single',
+  },
+];
+
+describe('Mask Snapshot', () => {
+  generateCanvasTestCases('Mask', TEST_CASES);
+});
